Move CleanupFixture to its own module.
